master

+++ resources/js/react/UI/comments/CommentsBody.jsx
@@ -1,5 +1,6 @@
 import React, { createContext, useContext, useEffect } from "react";
 import style from "./styles/comments.module.css";
+import './editor.style.css';
 import { AiOutlineFullscreen, AiOutlineFullscreenExit } from "react-icons/ai";
 import {
     IoIosArrowDown,
@@ -36,6 +37,10 @@ import getTextContent, {
 } from "./utils/getTextContent";
 import { useDeleteCommentsMutation } from "../../services/api/commentsApiSlice";
 import { useParams } from "react-router-dom";
+import ImageSliderModal from "./components/ImageSliderModal";
+import { isHasPermissionForWriteComment } from "./utils/isHasPermissionForWriteComment";
+import Sendbox from "./components/sendbox/Sendbox";
+
 
 const CommentContext = createContext({
     setScroll: () => {},
@@ -45,6 +50,11 @@ const CommentContext = createContext({
     setMentionedComment: () => {},
     contextHolder: {},
     setContextHolder: () => {},
+    allComments: [],
+    isImageModalOpen: false,
+    setIsImageModalOpen: () => {},
+    imageModalCurrentFileUrl: "",
+    setImageModalCurrentFileUrl: () => {},
 });
 export function useCommentContext() {
     return useContext(CommentContext);
@@ -60,7 +70,13 @@ const CommentsBody = ({
     fetching,
     refetch,
     taskId,
+    task,
     height,
+    showFullScreenBtn = true,
+    showCloseBtn = true,
+    showCommentEditor = true,
+    showSearchBtn = true,
+    onSubmit = async () => null,
 }) => {
     const param = useParams();
     const [deleteComments, { isLoading: deleteLoading }] =
@@ -81,7 +97,11 @@ const CommentsBody = ({
     const [selectedComments, setSecletedComments] = useState({});
     const [mentionedComment, setMentionedComment] = useState(null);
     const [contextHolder, setContextHolder] = useState(null);
+    const [isImageModalOpen, setIsImageModalOpen] = useState(false);
+    const [imageModalCurrentFileUrl, setImageModalCurrentFileUrl] =
+        useState("");
     // =================================================================
+ 
 
     useEffect(() => {
         setAllComments(comments);
@@ -200,7 +220,7 @@ const CommentsBody = ({
                 .getElementById(
                     searchIndexes[searchIndexes.length - commentIndex]
                 )
-                .scrollIntoView({
+                ?.scrollIntoView({
                     behavior: "smooth",
                     // block: "",
                 });
@@ -213,8 +233,8 @@ const CommentsBody = ({
     // scroll to the mention comment according to selection
     useEffect(() => {
         let time_ref;
-        if (selectMentionIndex) {
-            document.getElementById(selectMentionIndex).scrollIntoView({
+        if (selectMentionIndex && param?.taskId) {
+            document.getElementById(selectMentionIndex)?.scrollIntoView({
                 behavior: "smooth",
                 // block: "",
             });
@@ -222,7 +242,11 @@ const CommentsBody = ({
                 setSelectMentionIndex(0);
             }, 2000);
         }
-        return () => clearTimeout(time_ref);
+        return () => {
+            if (time_ref) {
+                clearTimeout(time_ref);
+            }
+        };
     }, [selectMentionIndex]);
 
     const handleCopyComments = () => {
@@ -283,29 +307,49 @@ const CommentsBody = ({
         // console.log({ commentsId,selectedComments });
         // return;
         // setIsLoading(true),
-        deleteComments({ commentsId })
-            .then(() => {
-                Swal.fire({
-                    icon: "success",
-                    title: "Comments deleted successfully",
-                    timer: 2000,
-                    showConfirmButton: true,
-                    timerProgressBar: true,
+        const deleteComments = async ({commentsId}) => {
+            await deleteComments({ commentsId })
+                .then(() => {
+                    Swal.fire({
+                        icon: "success",
+                        title: "Comments deleted successfully",
+                        timer: 2000,
+                        showConfirmButton: true,
+                        timerProgressBar: true,
+                    });
+                    setSecletedComments({});
+                })
+                .catch(() => {
+                    Swal.fire({
+                        icon: "error",
+                        title: "An error occured to delete comments",
+                        timer: 2000,
+                        showConfirmButton: true,
+                        timerProgressBar: true,
+                    });
                 });
-                setSecletedComments({});
-            })
-            .catch(() => {
-                Swal.fire({
-                    icon: "error",
-                    title: "An error occured to delete comments",
-                    timer: 2000,
-                    showConfirmButton: true,
-                    timerProgressBar: true,
-                });
-            });
+        }
         // setScroll(prev=>!prev);
         // refetch();
         // setIsLoading(false);
+
+
+        Swal.fire({
+            icon: 'warning',
+            title: 'Are you sure?',
+            text: 'If you click on Yes, the comment will be deleted. If you click on no, the comment will not be deleted.',
+            showConfirmButton: true,
+            confirmButtonText: 'Yes',
+            showDenyButton: true,
+            denyButtonText: 'No',
+            customClass: {
+                confirmButton: 'btn btn-primary',
+            }
+        }).then(res => {
+            if(res.isConfirmed){
+                deleteComments({commentsId})
+            }
+        })
     };
 
     const handleCopySingleComment = (comment) => {
@@ -351,7 +395,8 @@ const CommentsBody = ({
     const handleDeleteSingleComment = (comment) => {
         // console.log({ id: comment.id });
         // return;
-        deleteComments({ commentsId: [comment.id] })
+        const deleteComment = async(commentId) => {
+            await deleteComments({ commentsId: [commentId] })
             .then((res) => {
                 if (res.data.status == 200) {
                     Swal.fire({
@@ -380,6 +425,25 @@ const CommentsBody = ({
                     timerProgressBar: true,
                 });
             });
+        }
+       
+
+            Swal.fire({
+                icon: 'warning',
+                title: 'Are you sure?',
+                text: 'If you click on Yes, the comment will be deleted. If you click on no, the comment will not be deleted.',
+                showConfirmButton: true,
+                confirmButtonText: 'Yes',
+                showDenyButton: true,
+                denyButtonText: 'No',
+                customClass: {
+                    confirmButton: 'btn btn-primary',
+                }
+            }).then(res => {
+                if(res.isConfirmed){
+                    deleteComment(comment.id);
+                }
+            })
     };
 
     // console.log({allComments});
@@ -393,6 +457,11 @@ const CommentsBody = ({
                 setMentionedComment,
                 contextHolder,
                 setContextHolder,
+                allComments,
+                isImageModalOpen,
+                setIsImageModalOpen,
+                imageModalCurrentFileUrl,
+                setImageModalCurrentFileUrl,
             }}
         >
             <div
@@ -400,9 +469,11 @@ const CommentsBody = ({
                 style={{
                     backgroundImage: `url(${commentBg})`,
                     // backgroundImage:`url(https://seopage1storage.s3.ap-southeast-1.amazonaws.com/655f048a34e53.jpg)`,
-                    width: fullScreenView ? "100vw" : "auto",
-                    height: fullScreenView ? "99vh" : height,
-                    maxHeight: fullScreenView ? "99vh" : "auto",
+                    // width: fullScreenView ? "100vw" : "auto",
+                    width: fullScreenView ? "100%" : "auto",
+                    // height: fullScreenView ? "99vh" : height,
+                    height: fullScreenView ? "100%" : height,
+                    // maxHeight: fullScreenView ? "99vh" : "auto",
                 }}
             >
                 <header className={style.commentsBody_header}>
@@ -552,7 +623,7 @@ const CommentsBody = ({
                     )}
 
                     {/* search btn */}
-                    {param?.taskId ? (
+                    {param?.taskId && showSearchBtn ? (
                         <span
                             onClick={() => {
                                 if (showSearchBar) {
@@ -586,7 +657,7 @@ const CommentsBody = ({
                     )}
 
                     {/* full screen btn */}
-                    {param?.taskId ? (
+                    {param?.taskId && showFullScreenBtn ? (
                         !fullScreenView ? (
                             <AiOutlineFullscreen
                                 onClick={() => setFullScreenView(true)}
@@ -603,72 +674,76 @@ const CommentsBody = ({
                     )}
 
                     {/* cancel btn */}
-                    <span
-                        onClick={() => {
-                            if (setFullScreenView) {
-                                setFullScreenView(false);
-                            }
-                            close();
-                        }}
-                        className={`${style.commentsBody_header_btn}`}
-                    >
-                        <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            xlink="http://www.w3.org/1999/xlink"
-                            viewBox="15 7 28 28"
+                    {showCloseBtn ? (
+                        <span
+                            onClick={() => {
+                                if (setFullScreenView) {
+                                    setFullScreenView(false);
+                                }
+                                close();
+                            }}
+                            className={`${style.commentsBody_header_btn}`}
                         >
-                            <defs>
-                                <filter
-                                    id="Ellipse_58"
-                                    x="0"
-                                    y="0"
-                                    width="58"
-                                    height="58"
-                                    filterUnits="userSpaceOnUse"
-                                >
-                                    <feOffset dy="8" input="SourceAlpha" />
-                                    <feGaussianBlur
-                                        stdDeviation="5"
-                                        result="blur"
-                                    />
-                                    <feFlood
-                                        floodColor="#757575"
-                                        floodOpacity="0.161"
-                                    />
-                                    <feComposite operator="in" in2="blur" />
-                                    <feComposite in="SourceGraphic" />
-                                </filter>
-                            </defs>
-                            <g
-                                id="Group_3644"
-                                data-name="Group 3644"
-                                transform="translate(-976 -145)"
+                            <svg
+                                xmlns="http://www.w3.org/2000/svg"
+                                xlink="http://www.w3.org/1999/xlink"
+                                viewBox="15 7 28 28"
                             >
+                                <defs>
+                                    <filter
+                                        id="Ellipse_58"
+                                        x="0"
+                                        y="0"
+                                        width="58"
+                                        height="58"
+                                        filterUnits="userSpaceOnUse"
+                                    >
+                                        <feOffset dy="8" input="SourceAlpha" />
+                                        <feGaussianBlur
+                                            stdDeviation="5"
+                                            result="blur"
+                                        />
+                                        <feFlood
+                                            floodColor="#757575"
+                                            floodOpacity="0.161"
+                                        />
+                                        <feComposite operator="in" in2="blur" />
+                                        <feComposite in="SourceGraphic" />
+                                    </filter>
+                                </defs>
                                 <g
-                                    transform="matrix(1, 0, 0, 1, 976, 145)"
-                                    filter="url(#Ellipse_58)"
+                                    id="Group_3644"
+                                    data-name="Group 3644"
+                                    transform="translate(-976 -145)"
                                 >
-                                    <circle
-                                        id="Ellipse_58-2"
-                                        data-name="Ellipse 58"
-                                        cx="14"
-                                        cy="14"
-                                        r="14"
-                                        transform="translate(15 7)"
-                                        fill="#df0b0b"
+                                    <g
+                                        transform="matrix(1, 0, 0, 1, 976, 145)"
+                                        filter="url(#Ellipse_58)"
+                                    >
+                                        <circle
+                                            id="Ellipse_58-2"
+                                            data-name="Ellipse 58"
+                                            cx="14"
+                                            cy="14"
+                                            r="14"
+                                            transform="translate(15 7)"
+                                            fill="#df0b0b"
+                                        />
+                                    </g>
+                                    <path
+                                        id="remove_1_"
+                                        data-name="remove (1)"
+                                        d="M5.059,13A7.941,7.941,0,1,1,13,20.941,7.941,7.941,0,0,1,5.059,13ZM13,4a9,9,0,0,0,0,18,9.139,9.139,0,0,0,6.911-3.235A8.762,8.762,0,0,0,22,13,9,9,0,0,0,13,4ZM9.256,15.995,12.251,13,9.256,10.005,10,9.257,13,12.251l3-3,.748.748-3,3,3,3-.748.748-3-2.995L10,16.743Z"
+                                        transform="translate(992 153)"
+                                        fill="#fff"
+                                        fillRule="evenodd"
                                     />
                                 </g>
-                                <path
-                                    id="remove_1_"
-                                    data-name="remove (1)"
-                                    d="M5.059,13A7.941,7.941,0,1,1,13,20.941,7.941,7.941,0,0,1,5.059,13ZM13,4a9,9,0,0,0,0,18,9.139,9.139,0,0,0,6.911-3.235A8.762,8.762,0,0,0,22,13,9,9,0,0,0,13,4ZM9.256,15.995,12.251,13,9.256,10.005,10,9.257,13,12.251l3-3,.748.748-3,3,3,3-.748.748-3-2.995L10,16.743Z"
-                                    transform="translate(992 153)"
-                                    fill="#fff"
-                                    fillRule="evenodd"
-                                />
-                            </g>
-                        </svg>
-                    </span>
+                            </svg>
+                        </span>
+                    ) : (
+                        <></>
+                    )}
                 </header>
 
                 <main
@@ -710,7 +785,7 @@ const CommentsBody = ({
                                     />
                                 );
                             })}
-                            {(fetching || isloading || deleteLoading) ? (
+                            {fetching || isloading || deleteLoading ? (
                                 <div className="d-flex justify-content-center mt-2">
                                     <div
                                         className="spinner-border"
@@ -733,10 +808,34 @@ const CommentsBody = ({
                         }}
                         ref={chatbottom_ref}
                     />
+                    <ImageSliderModal
+                        isOpen={isImageModalOpen}
+                        close={() => setIsImageModalOpen(false)}
+                        selectedImgUrl={imageModalCurrentFileUrl}
+                        setSelectedImgUrl={setImageModalCurrentFileUrl}
+                    />
                 </main>
-                <footer className={`${style.commentsBody_inputField}`}>
-                    <ChatInput taskId={taskId} setScroll={setScroll} setIsLoading={setIsLoading} />
-                </footer>
+                 {isHasPermissionForWriteComment({
+                     assignTo: task?.assigneeTo?.id, 
+                     assignBy: task?.assigneeBy?.id
+                }) ?  
+                    <footer className={`${style.commentsBody_inputField}`}>
+                        <Sendbox 
+                            onSubmit={onSubmit}
+                            taskId={taskId}
+                            setScroll={setScroll}
+                            setIsLoading={setIsLoading}
+                        />
+
+                        {/* customize start from there */}
+                        {/* <ChatInput
+                            onSubmit={onSubmit}
+                            taskId={taskId}
+                            setScroll={setScroll}
+                            setIsLoading={setIsLoading}
+                        /> */}
+                    </footer>
+                 : null} 
 
                 {Object.keys(selectedComments).length > 0 ? (
                     <div
